Fix bug and clean up tests
