+++ cosmos-integration-tests/src/main/scala/com/mesosphere/cosmos/PackageListIntegrationSpec.scala
@@ -1,6 +1,5 @@
 package com.mesosphere.cosmos
 
-import cats.syntax.either._
 import com.mesosphere.cosmos.http.CosmosRequests
 import com.mesosphere.cosmos.repository.DefaultRepositories
 import com.mesosphere.cosmos.rpc.v1.circe.Decoders._
@@ -113,28 +112,22 @@ final class PackageListIntegrationSpec
   }
 
   private[this] def withDeletedRepository(repository: PackageRepository)(action: => Any): Any = {
-    val repoDeleteRequest = PackageRepositoryDeleteRequest(name = Some(repository.name))
-    val request = CosmosRequests.packageRepositoryDelete(repoDeleteRequest)
-    val actualDelete = CosmosClient.callEndpoint[PackageRepositoryDeleteResponse](request)
+
+    val originalRepositories = ItUtil.listRepositories()
+      .withClue("when getting original repositories")
+
+    val actualDelete = ItUtil.deleteRepository(repository)
       .withClue("when deleting repo")
 
     try {
-      assertResult(Right(None)) {
-        actualDelete.map(_.repositories.find(_.name == repository.name))
+      assertResult(None) {
+        actualDelete.repositories.find(_.name == repository.name)
       }
 
       action
     } finally {
-      val repoAddRequest = PackageRepositoryAddRequest(repository.name, repository.uri)
-      val request = CosmosRequests.packageRepositoryAdd(repoAddRequest)
-      val actualAdd = CosmosClient.callEndpoint[PackageRepositoryAddResponse](request)
+      ItUtil.replaceRepositoriesWith(originalRepositories)
         .withClue("when restoring deleted repo")
-
-      val _ = inside(actualAdd) { case Right(PackageRepositoryAddResponse(repositories)) =>
-        inside(repositories.find(_.name == repository.name)) { case Some(addedRepository) =>
-          assertResult(repository)(addedRepository)
-        }
-      }
     }
   }
