+++ cosmos-integration-tests/src/main/scala/com/mesosphere/cosmos/PackageRepositoryIntegrationSpec.scala
@@ -1,9 +1,9 @@
 package com.mesosphere.cosmos
 
+import com.mesosphere.cosmos.ItUtil._
 import com.mesosphere.cosmos.http.CosmosRequests
 import com.mesosphere.cosmos.repository.DefaultRepositories
 import com.mesosphere.cosmos.repository.PackageRepositorySpec
-import com.mesosphere.cosmos.rpc.v1.circe.Decoders._
 import com.mesosphere.cosmos.rpc.v1.model._
 import com.mesosphere.cosmos.test.CosmosIntegrationTestClient.CosmosClient
 import com.netaporter.uri.dsl._
@@ -12,7 +12,6 @@ import org.scalatest.BeforeAndAfter
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.FreeSpec
 import org.scalatest.prop.TableDrivenPropertyChecks
-import scala.util.Right
 
 final class PackageRepositoryIntegrationSpec extends FreeSpec with BeforeAndAfter with BeforeAndAfterAll {
 
@@ -110,44 +109,11 @@ final class PackageRepositoryIntegrationSpec extends FreeSpec with BeforeAndAfte
 
 private object PackageRepositoryIntegrationSpec extends TableDrivenPropertyChecks {
 
-  private def listRepositories(): Seq[PackageRepository] = {
-    val request = CosmosRequests.packageRepositoryList
-    val Right(response) = CosmosClient.callEndpoint[PackageRepositoryListResponse](request)
-    response.repositories
-  }
-
-  private def addRepository(
-    source: PackageRepository
-  ): PackageRepositoryAddResponse = {
-    val repoAddRequest = PackageRepositoryAddRequest(source.name, source.uri)
-    val request = CosmosRequests.packageRepositoryAdd(repoAddRequest)
-    val Right(response) = CosmosClient.callEndpoint[PackageRepositoryAddResponse](request)
-    response
-  }
-
-  private def deleteRepository(
-    source: PackageRepository
-  ): PackageRepositoryDeleteResponse = {
-    val repoDeleteRequest = PackageRepositoryDeleteRequest(name = Some(source.name))
-    val request = CosmosRequests.packageRepositoryDelete(repoDeleteRequest)
-    val Right(response) = CosmosClient.callEndpoint[PackageRepositoryDeleteResponse](request)
-    response
-  }
-
   private def sendAddRequest(
     addRequest: PackageRepositoryAddRequest
   ): Response = {
     val request = CosmosRequests.packageRepositoryAdd(addRequest)
     CosmosClient.submit(request)
-  }
-
-  private def replaceRepositoriesWith(repositories: Seq[PackageRepository]): Unit = {
-    listRepositories().foreach { repo =>
-      deleteRepository(repo)
-    }
-    repositories.foreach { repo =>
-      addRepository(repo)
-    }
   }
 
 }
